services/auth-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic-settings==2.2.1